BOAC-5877, move front-end type defs into 'lib/types.ts'; modularize EditUser dialog
