+++ src/components/note/AdvisingNote.vue
@@ -141,13 +141,14 @@
 import AdvisingNoteAttachments from '@/components/note/AdvisingNoteAttachments'
 import AdvisingNoteTopics from '@/components/note/AdvisingNoteTopics'
 import AreYouSureModal from '@/components/util/AreYouSureModal'
+import {DateTime} from 'luxon'
 import {addAttachments, removeAttachment} from '@/api/notes'
 import {alertScreenReader, numFormat, oxfordJoin, toInt} from '@/lib/utils'
 import {computed, onMounted, ref, watch} from 'vue'
 import {get, isNil, isNumber, map, orderBy, size} from 'lodash'
-import {DateTime} from 'luxon'
-import {getBoaUserRoles, termNameForSisId} from '@/berkeley'
+import {getBoaUserRoles} from '@/lib/berkeley-department'
 import {getCalnetProfileByCsid, getCalnetProfileByUid} from '@/api/user'
+import {termNameForSisId} from '@/berkeley'
 import {useContextStore} from '@/stores/context'
 import {useNoteStore} from '@/stores/note-edit-session'
