+++ src/lib/utils.ts
@@ -1,123 +1,9 @@
 import numeral from 'numeral'
-import {Cohort, CuratedGroup} from '@/lib/cohort'
+import {BoaUser, Cohort, CuratedGroup, ScreenReaderAlert} from '@/lib/types'
 import {concat, head, initial, isNil, isNumber, join, last, toLower, trim} from 'lodash'
 import {getUserProfile} from '@/api/user'
 import {nextTick} from 'vue'
 import {useContextStore} from '@/stores/context'
-
-export type BoaConfig = {
-  academicStandingDescriptions: object,
-  apiBaseUrl: string,
-  currentEnrollmentTerm: string,
-  currentEnrollmentTermId: number,
-  defaultTermUnitsAllowed: {
-    max: number,
-    min: number
-  },
-  devAuthEnabled: boolean,
-  draftNoteSubjectPlaceholder: string,
-  fixedWarningOnAllPages: boolean,
-  gaMeasurementId: string,
-  isProduction: boolean,
-  isVueAppDebugMode: boolean,
-  maxAttachmentsPerNote: number,
-  notesDraftAutoSaveInterval: number,
-  supportEmailAddress: string,
-  timezone: string
-}
-
-export type Course = {
-  sections: Section[],
-  waitlisted: boolean
-}
-
-export type BoaUser = {
-  id: number | undefined,
-  automateDegreeProgressPermission: boolean,
-  canAccessAdmittedStudents: boolean,
-  canAccessAdvisingData: boolean,
-  canAccessCanvasData: boolean,
-  canEditDegreeProgress: boolean,
-  canReadDegreeProgress: boolean,
-  degreeProgressPermission: string | undefined,
-  deletedAt: Date | undefined,
-  departments: BoaUserDepartment[],
-  inDemoMode: boolean,
-  isAdmin: boolean,
-  isAuthenticated: boolean,
-  isBlocked: boolean,
-  isDemoModeAvailable: boolean,
-  myCohorts: Cohort[],
-  myCuratedGroups: CuratedGroup[],
-  myDraftNoteCount: number | undefined,
-  name: string | undefined,
-  preferences: {
-    termId: string | undefined
-  },
-  title: string | undefined,
-  uid: string | undefined
-}
-
-export type BoaUserDepartment = {
-  id: number,
-  deptCode: string,
-  deptName: string,
-  memberships: DepartmentMembership[]
-}
-
-export type Department = {
-  id: number,
-  deptCode: string,
-  deptName: string,
-  peerAdvisingDepartments: PeerAdvisingDepartment[]
-}
-
-export type DepartmentMembership = {
-  automateMembership?: boolean,
-  peerAdvisingDepartmentId?: number,
-  peerAdvisingDepartmentName?: string,
-  role: DepartmentMembershipRole | undefined
-}
-
-export type DepartmentMembershipRole = 'advisor' | 'director' | 'peer_advisor' | 'peer_advisor_manager'
-
-export type ExportListOption = {
-  text: string,
-  value: string,
-  disabled?: boolean
-}
-
-export type Pagination = {
-  currentPage: number,
-  itemsPerPage: number
-}
-
-export type PeerAdvisingDepartment = {
-  'id': number,
-  'name': string
-}
-
-export type ScreenReaderAlert = {
-  message: string,
-  politeness: string
-}
-
-export type Section = {
-  enrollmentStatus: string,
-  gradingBasis: string,
-  incompleteComments: string,
-  incompleteLapseGradeDate: string,
-  incompleteStatusCode: string
-}
-
-export type ServiceAnnouncement = {
-  isPublished: boolean,
-  text: string
-}
-
-export type Student = {
-  sid: string
-}
 
 export const ANONYMOUS_USER: BoaUser = {
   id: undefined,
