+++ src/router.ts
@@ -1,5 +1,3 @@
-import PeerAdvisor from '@/views/PeerAdvisor.vue'
-
 const AdmitStudent = () => import('@/views/AdmitStudent.vue')
 const AdmitStudents = () => import('@/views/AdmitStudents.vue')
 const AllCohorts = () => import('@/views/AllCohorts.vue')
@@ -14,13 +12,14 @@ const DegreeTemplate = () => import('@/views/degree/DegreeTemplate.vue')
 const DraftNotes = () => import('@/views/DraftNotes.vue')
 const Error = () => import('@/views/Error.vue')
 const FlightDeck = () => import('@/views/FlightDeck.vue')
-const Home = () => import('@/views/Home.vue')
 const FlightDataRecorder = () => import('@/views/FlightDataRecorder.vue')
+const Home = () => import('@/views/Home.vue')
 const Login = () => import('./layouts/Login.vue')
 const ManageDegreeChecks = () => import('@/views/degree/ManageDegreeChecks.vue')
 const NotFound = () => import('@/views/NotFound.vue')
 const PassengerManifest = () => import('@/views/PassengerManifest.vue')
 const PeerAdvisingLayout = () => import('@/layouts/PeerAdvisingLayout.vue')
+import PeerAdvisor from '@/views/PeerAdvisor.vue'
 const PeerAdvisorManager = () => import('@/views/PeerAdvisorManager.vue')
 const PrintableDegreeTemplate = () => import('@/views/degree/PrintableDegreeTemplate.vue')
 const Profile = () => import('@/views/Profile.vue')
@@ -30,12 +29,12 @@ const StandardLayout = () => import('@/layouts/StandardLayout.vue')
 const StudentDegreeCheck = () => import('@/views/degree/StudentDegreeCheck.vue')
 const StudentDegreeCreate = () => import('@/views/degree/StudentDegreeCreate.vue')
 const StudentDegreeHistory = () => import('@/views/degree/StudentDegreeHistory.vue')
-import {BoaUser} from './lib/utils'
+import {BoaUser} from './lib/types'
 import {NavigationGuardNext, RouteLocation, RouteRecordRaw, createRouter, createWebHistory} from 'vue-router'
 import {filter, get, includes, size, toString, trim} from 'lodash'
-import {isAdvisor, isDirector, isPeerAdvisor, isPeerAdvisorManager} from '@/berkeley'
 import {useContextStore} from '@/stores/context'
 import {useSearchStore} from '@/stores/search'
+import {isAdvisor, isDirector, isPeerAdvisor, isPeerAdvisorManager} from '@/lib/boa-user'
 
 const $_goToLogin = (to: RouteLocation, next: NavigationGuardNext) => {
   next({
